More readable concurrent team matches check
Using a flatMap might optimize the performance a bit as well, but I'm leaving the comments anyway.
Would likely point it out as a topic for discussion in code review.

Team names should be Enums as well to lower the risk of spelling errors.
Too much work for this type of assignment, but something to consider for production ready code.