Second summary iteration
It felt clumpsy to do the sorting on both key and value in an entryset, so I tried switching to just a List for matches with Match owning its own id.
This also adheres a better to OO-principles.

Since there is no Map-specific algorithms taking advantage of the key/value structure I don't think it's a performance loss.

But the simplified sorting-syntax does improve readability.
With a slight adjustment for getting matches with id.