MatchSummary record for summary data
Internally the match data is logically structured based on OO-principles,
but that's not really relevant for the summary according to the specifications.
There a more flat structure and without match id is expected.

Therefore, a MatchSummary record has been added which the internal Match data kan be mapped to.

This also provides some basic separation of concerns with an internal data model as well as a presentation model of sorts.