Finish match fix
Switching from Map to List gave remove() a different behaviour,
although still compatible.
Since the matchId is currently the equivalent of an incrementing index,
the index-based remove() still worked.
But that was more by accident than by design,
so I updated the implementation to not be dependent on it.
It now uses matchId as designed.