Improved multi match test and fixed bug in updateMatch
Decided that the big scoreboard test should also finish a match and add a new with re-used team names.

Doing that exposed a similar bug in updateMatch as previously fixed in getMatch.

Decided to simplify the implementation to let it throw an exception directly as there is no fallback logic anyway.